Extract audit event logic
